requirements.txt
flask==2.3.3
flask-socketio==5.3.6
gpiozero==2.0.1
smbus2==0.4.3
numpy==1.26.0
opencv-python==4.8.1.78
aiortc==1.5.0
aiohttp==3.9.1
whisper==1.1.10
requests==2.31.0
python-dotenv==1.0.0
RPi.GPIO==0.7.1
picamera2==0.3.12
pydantic==2.5.2
websockets==11.0.3 